pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "open-fdd"
version = "2.0.2"
description = "Fault Detection and Diagnostics for HVAC systems — config-driven, pandas-based"
readme = "README.md"
license = { text = "MIT" }
authors = [
    { name = "Ben Bartling", email = "ben.bartling@gmail.com" }
]
requires-python = ">=3.9"
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "Intended Audience :: Science/Research",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Programming Language :: Python :: 3.14",
    "Topic :: Scientific/Engineering",
]
keywords = ["hvac", "fdd", "fault-detection", "building-automation", "pandas", "brick"]

dependencies = [
    "pandas",
    "pyyaml",
]

[project.optional-dependencies]
# Development: use this for clone + run tests + black. pip install -e ".[dev]"
dev = [
    "aiohttp",
    "pytest",
    "black[jupyter]",
    "pre-commit",
    "httpx",
    "psycopg2-binary",
    "requests",
    "pydantic-settings",
    "fastapi",
    "uvicorn[standard]",
    "python-multipart",
    "rdflib",
]
# Same as dev (CI and docs may use .[test]; prefer .[dev] for local development)
test = [
    "aiohttp",
    "pytest",
    "black[jupyter]",
    "pre-commit",
    "httpx",
    "psycopg2-binary",
    "requests",
    "pydantic-settings",
    "fastapi",
    "uvicorn[standard]",
    "python-multipart",
    "rdflib",
]
viz = [
    "matplotlib",
]
brick = [
    "rdflib",
]
docx = [
    "python-docx",
]
platform = [
    "psycopg2-binary",
    "requests",
    "httpx",
    "pydantic-settings",
    "fastapi",
    "uvicorn[standard]",
    "python-multipart",
]
bacnet = [
    "bacpypes3",
    "ifaddr",
    "httpx",
]
monitoring = [
    "docker",
]

[project.urls]
Homepage = "https://github.com/bbartling/open-fdd"
Documentation = "https://bbartling.github.io/open-fdd/"
Repository = "https://github.com/bbartling/open-fdd"

[tool.setuptools.packages.find]
include = ["open_fdd*"]

[tool.setuptools.package-data]
open_fdd = ["rules/*.yaml", "platform/static/*"]

[tool.black]
line-length = 88
target-version = ["py314"]
# Exclude files that trigger Black bugs or are intentionally unformatted (e2e script).
extend-exclude = "open_fdd/platform/data_model_ttl\\.py|tools/graph_and_crud_test\\.py"

[tool.pytest.ini_options]
testpaths = ["open_fdd/tests", "stack/ha_integration/tests"]
pythonpath = ["."]
filterwarnings = [
    "ignore::DeprecationWarning:rdflib",
    "ignore::DeprecationWarning:pyparsing",
]
